{{cookiecutter.project_slug}}/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
_py-xgboost-mutex=2.0=cpu_0
alembic=1.8.1=pypi_0
aom=3.4.0=hbe31e94_1
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39hb18efdd_2
asttokens=2.0.5=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
autopage=0.5.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.1=pyhd8ed1ab_0
blosc=1.21.1=hd414afc_3
boost-cpp=1.74.0=h1cb353e_8
brotli=1.0.9=h1c322ee_7
brotli-bin=1.0.9=h1c322ee_7
brotlipy=0.7.0=py39hb18efdd_1004
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.6.15=h4653dfc_0
cairo=1.16.0=h3e596be_1011
certifi=2022.6.15=py39h2804cbe_0
cffi=1.15.1=py39h04d3946_0
cfitsio=4.1.0=hd4f5c17_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
click=8.1.3=py39h2804cbe_0
click-plugins=1.1.1=py_0
cliff=3.10.1=pypi_0
cligj=0.7.2=pyhd8ed1ab_1
cmaes=0.8.2=pypi_0
cmd2=2.4.2=pypi_0
colorlog=6.6.0=pypi_0
cryptography=37.0.4=py39h4cb6239_0
curl=7.83.1=h2fcd78c_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.0=py39h0ef5a74_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.9.1=pyhd8ed1ab_0
expat=2.4.8=h6b3803e_0
ffmpeg=4.4.2=gpl_h0b00469_106
fiona=1.8.21=py39hf45f784_2
flit-core=3.7.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=hfb34624_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.34.4=py39h9eb174b_0
freetype=2.10.4=h17b34a0_1
freexl=1.0.6=h27ca646_0
gdal=3.5.1=py39haf2729a_2
geopandas=0.9.0=pyhd8ed1ab_0
geos=3.11.0=h9a09cb3_0
geotiff=1.7.1=hc898e3f_3
gettext=0.19.8.1=h049c9fb_1008
giflib=5.2.1=h27ca646_2
gmp=6.2.1=h9f76cd9_0
gnutls=3.7.6=h27a1928_5
graphite2=1.3.13=h9f76cd9_1001
harfbuzz=5.1.0=hd36a07e_0
hdf4=4.2.15=h0ef36ac_3
hdf5=1.12.2=nompi_h8968d4b_100
icu=70.1=h6b3803e_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.4=py39h2804cbe_0
importlib_resources=5.9.0=pyhd8ed1ab_0
ipykernel=6.15.1=pyh736e0ef_0
ipython=8.4.0=py39h2804cbe_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.1=pyhd8ed1ab_0
jasper=2.0.33=hba35424_0
jedi=0.18.1=py39h2804cbe_1
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=he4db4b2_2
json-c=0.16=hc449e50_0
jsonschema=4.9.1=pyhd8ed1ab_0
jupyter=1.0.0=py39h2804cbe_7
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_console=6.4.4=pyhd8ed1ab_0
jupyter_core=4.11.1=py39h2804cbe_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=1.1.1=pyhd8ed1ab_0
kealib=1.4.15=h02ce806_1
kiwisolver=1.4.4=py39hab5e169_0
krb5=1.19.3=hf9b2bbe_0
lame=3.100=h27ca646_1001
lcms2=2.12=had6a04f_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=15_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=15_osxarm64_openblas
libcurl=7.83.1=h2fcd78c_0
libcxx=14.0.6=h04bba0f_0
libdap4=3.20.6=h8510809_2
libdeflate=1.13=h57fd34a_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgdal=3.5.1=hc0c25a4_2
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
libglib=2.72.1=ha1047ec_0
libiconv=1.16=h642e427_0
libidn2=2.3.3=he4db4b2_0
libkml=1.3.0=h893ab4d_1014
liblapack=3.9.0=15_osxarm64_openblas
liblapacke=3.9.0=15_osxarm64_openblas
libnetcdf=4.8.1=nompi_hc076e8b_103
libnghttp2=1.47.0=he723fca_0
libopenblas=0.3.20=openmp_h130de29_1
libopencv=4.6.0=py39he1c1adf_3
libpng=1.6.37=h2c9beb0_3
libpq=14.4=h8ab49ba_0
libprotobuf=3.20.1=h332123e_0
librttopo=1.1.0=h275bb25_11
libsodium=1.0.18=h27ca646_1
libspatialindex=1.9.3=hbdafb3b_4
libspatialite=5.0.1=h47b1232_18
libssh2=1.10.0=hb80f160_2
libtasn1=4.18.0=he4db4b2_1
libtiff=4.4.0=hd6ca8d7_3
libunistring=0.9.10=h3422bc3_0
libvpx=1.11.0=hc470f4d_3
libwebp=1.2.3=h1e26fe3_1
libwebp-base=1.2.3=he4db4b2_2
libxcb=1.13=h9b22ae9_1004
libxgboost=1.5.1=cpu_h9f176f6_2
libxml2=2.9.14=h035c1df_3
libzip=1.9.2=h8caee39_0
libzlib=1.2.12=ha287fd2_2
lightgbm=3.3.2=py39h8685b90_0
llvm-openmp=14.0.4=hd125106_0
lz4-c=1.9.3=hbdafb3b_1
mako=1.2.1=pypi_0
markupsafe=2.1.1=py39hb18efdd_1
matplotlib=3.5.2=py39hdf13c20_1
matplotlib-base=3.5.2=py39h4ee150f_1
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py39h5161555_1005
munch=2.5.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.5=py39h2804cbe_0
nbformat=5.4.0=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
nettle=3.8.1=h63371fa_1
notebook=6.4.12=pyha770c72_0
nspr=4.32=hbdafb3b_1
nss=3.78=h1483a63_0
numpy=1.23.1=py39h7df2422_0
opencv=4.6.0=py39hdf13c20_3
openh264=2.2.0=h9a09cb3_2
openjpeg=2.4.0=h062765e_1
openssl=1.1.1q=ha287fd2_0
optuna=2.10.1=pypi_0
p11-kit=0.24.1=h29577a5_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=py39hd2dba81_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.2=pyhd8ed1ab_0
pbr=5.9.0=pypi_0
pcre=8.45=hbdafb3b_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=py39h1b8be2f_0
pip=22.2.2=pyhd8ed1ab_0
pixman=0.40.0=h27ca646_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
poppler=22.04.0=hadf1f10_1
poppler-data=0.4.11=hd8ed1ab_0
postgresql=14.4=heaa1073_0
prettytable=3.3.0=pypi_0
proj=9.0.1=h4c79c2b_1
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.30=pyha770c72_0
prompt_toolkit=3.0.30=hd8ed1ab_0
psutil=5.9.1=py39h9eb174b_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-opencv=4.6.0=py39hfa6204d_3
py-xgboost=1.5.1=cpu_py39h6861412_2
pycparser=2.21=pyhd8ed1ab_0
pygments=2.12.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyperclip=1.8.2=pypi_0
pyproj=3.3.1=py39h33e860d_1
pyrsistent=0.18.1=py39hb18efdd_1
pysocks=1.7.1=py39h2804cbe_5
python=3.9.13=hc596b02_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.20.0=pyhd8ed1ab_0
python-fastjsonschema=2.16.1=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2022.1=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py39h8faa4b9_0
readline=8.1.2=h46ed386_0
requests=2.28.1=pyhd8ed1ab_0
rtree=1.0.0=py39hb28b0e7_1
scikit-learn=1.1.2=py39h598ef33_0
scipy=1.9.0=py39h14896cb_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-analyzer=0.2.11=pypi_0
seaborn-base=0.11.2=pyhd8ed1ab_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=63.4.2=py39h2804cbe_0
shapely=1.8.2=py39h12552bf_3
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h39c3846_1
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlalchemy=1.4.39=pypi_0
sqlite=3.39.2=h40dfcc0_0
stack_data=0.3.0=pyhd8ed1ab_0
statsmodels=0.13.2=py39h7b9fbcb_0
stevedore=4.0.0=pypi_0
svt-av1=1.1.0=hbe31e94_1
terminado=0.15.0=py39h2804cbe_0
testpath=0.6.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tiledb=2.9.4=hdd33e7b_0
tk=8.6.12=he1e0b03_0
tornado=6.2=py39h9eb174b_0
tqdm=4.64.0=pypi_0
traitlets=5.3.0=pyhd8ed1ab_0
tzcode=2022a=h1c322ee_0
tzdata=2022a=h191b570_0
unicodedata2=14.0.0=py39hb18efdd_1
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.1=pyha770c72_0
x264=1!161.3030=h3422bc3_1
x265=3.5=hbc6ce65_3
xerces-c=3.2.3=h1b0b642_5
xgboost=1.5.1=cpu_py39h5674d11_2
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
zeromq=4.3.4=hbdafb3b_1
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=ha287fd2_2
zstd=1.5.2=hd705a24_2
